fix: change datatype to integer for id

+++ lostandfound/src/main/java/com/example/lostandfound/controller/AdminController.java
@@ -1,13 +1,19 @@
 package com.example.lostandfound.controller;
 
-import com.example.lostandfound.model.AdminAction;
-import com.example.lostandfound.model.User;
-import com.example.lostandfound.model.Item;
-import com.example.lostandfound.service.AdminService;
-import org.springframework.beans.factory.annotation.Autowired;
-import org.springframework.web.bind.annotation.*;
-
 import java.util.List;
+
+import org.springframework.beans.factory.annotation.Autowired;
+import org.springframework.web.bind.annotation.GetMapping;
+import org.springframework.web.bind.annotation.PostMapping;
+import org.springframework.web.bind.annotation.RequestBody;
+import org.springframework.web.bind.annotation.RequestMapping;
+import org.springframework.web.bind.annotation.RequestParam;
+import org.springframework.web.bind.annotation.RestController;
+
+import com.example.lostandfound.model.AdminAction;
+import com.example.lostandfound.model.Item;
+import com.example.lostandfound.model.User;
+import com.example.lostandfound.service.AdminService;
 
 @RestController
 @RequestMapping("/api/admin")
@@ -30,7 +36,7 @@ public class AdminController {
 
     // Review and approve an item return
     @PostMapping("/approveReturn")
-    public String approveItemReturn(@RequestParam Long itemId, @RequestParam Integer userId) {
+    public String approveItemReturn(@RequestParam Integer itemId, @RequestParam Integer userId) {
         adminService.approveItemReturn(itemId, userId);
         return "Item return approved, merit points updated.";
     }
